refactor(naming): proper hook names
